% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulations.R
\name{changeFreqs}
\alias{changeFreqs}
\title{Change to minor allele frequency}
\usage{
changeFreqs(freqs)
}
\arguments{
\item{freqs}{a vector of allele frequencies where each entry corresponds to a
different site.}
}
\value{
a vector with the same length as \code{freqs} with the allele
frequencies above 0.5 replaced by the minor allele frequency.
}
\description{
Ensures that the allele frequencies computed directly from genotypes
correspond to the minor allele frequencies at each site.
}
\examples{
set.seed(10)
freqs <- runif(20)
changeFreqs(freqs)

}
