% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pool.R
\name{ExpHet_site}
\alias{ExpHet_site}
\title{Compute expected heterozygosity per site}
\usage{
ExpHet_site(geno_site)
}
\arguments{
\item{geno_site}{is a vector where each entry contains the genotype for a
single individual, coded as 0,1,2 and using NA for the missing data.}
}
\value{
a numerical value corresponding to the expected heterozygosity at
that site.
}
\description{
Computes the expected heterozygosity for a given site.
}
\keyword{internal}
