% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pool.R
\name{Pfreqs}
\alias{Pfreqs}
\title{Compute allele frequencies from pooled sequencing data}
\usage{
Pfreqs(minor, coverage, min.minor, ifreqs)
}
\arguments{
\item{minor}{a matrix with the number of minor-allele reads. Each row should
be a different population and each column a different site.}

\item{coverage}{a matrix with the total coverage. Each row should be a
different population and each column a different site.}

\item{min.minor}{is an integer representing the minimum allowed number of
minor-allele reads. Sites that, across all populations, have less
minor-allele reads than this threshold will be removed from the data.}

\item{ifreqs}{a vector of allele frequencies computed directly from the
genotypes where each entry corresponds to a different site.}
}
\value{
a list with two entries. The \code{ifreqs} entry contains the allele
frequencies computed directly from genotypes and \code{pfreqs} the allele
frequencies computed from pooled sequencing data.
}
\description{
Computes the minor-allele frequency from pooled data and removes any site
with too few minor-allele reads from both the pool frequencies and
frequencies computed directly from genotypes.
}
\details{
The frequency at a given SNP is calculated according to: π = c/r, where c =
number of minor-allele reads and r = total number of observed reads.
Additionally, if a site has less minor-allele reads than \code{min.minor}
across all populations, that site is removed from the data.
}
\examples{
set.seed(10)
freqs <- runif(20)
freqs <- changeFreqs(freqs)
set.seed(10)
minor <- matrix(sample(x = c(0,5,10), size = 20, replace = TRUE), nrow = 1)
coverage <- matrix(sample(100:150, size = 20), nrow = 1)
Pfreqs(minor = minor, coverage = coverage, min.minor = 2, ifreqs = freqs)

}
