% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulations.R
\name{removeSites}
\alias{removeSites}
\title{Apply a minor allele reads threshold}
\usage{
removeSites(freqs, minor, coverage, min.minor)
}
\arguments{
\item{freqs}{a vector of allele frequencies where each entry corresponds to a
different site.}

\item{minor}{a matrix with the number of minor-allele reads. Each row should
be a different population and each column a different site.}

\item{coverage}{a matrix with the total coverage. Each row should be a
different population and each column a different site.}

\item{min.minor}{is an integer representing the minimum allowed number of
minor-allele reads. Sites that, across all populations, have less
minor-allele reads than this threshold will be removed from the data.}
}
\value{
a list with three named entries. \code{freqs} contains the allele
frequencies minus the frequency of any removed site. \code{minor} contains
the number of minor-allele reads minus any removed site and the
\code{coverage} entry contains the total coverage minus any removed site.
}
\description{
Removes sites where the total number of minor-allele reads is below a certain
threshold.
}
\details{
If a site has less minor-allele reads than \code{min.minor} across all
populations, that site is removed from the data.
}
\examples{
freqs <- runif(20)
set.seed(10)
minor <- matrix(sample(x = c(0,5,10), size = 20, replace = TRUE), nrow = 1)
coverage <- matrix(sample(100:150, size = 20), nrow = 1)
removeSites(freqs = freqs, minor = minor, coverage, min.minor = 2)

}
