% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pool.R
\name{strg2sync}
\alias{strg2sync}
\title{Create sync string for a single SNP}
\usage{
strg2sync(reference, alternative)
}
\arguments{
\item{reference}{an integer representing the number of reads with the
reference allele.}

\item{alternative}{an integer representing the number of reads with the
alternative allele.}
}
\value{
a character string coded as
A-count:T-count:C-count:G-count:N-count:deletion-count.
}
\description{
Creates a string with the information for a single SNP. The information is
coded as A-count:T-count:C-count:G-count:N-count:deletion-count. Note that we
assume that the reference allele is always A and the alternative is always T.
}
\keyword{internal}
